moved ContextMenuExtensions into ContextMenu namespace
