+++ GitUI/CommandsDialogs/UserScriptContextMenuExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
-using GitUI.BranchTreePanel;
+using GitUI.BranchTreePanel.ContextMenu;
 using GitUI.Hotkey;
 using GitUI.Script;
 using ResourceManager;
